requirements.txt
Flask
Frozen-Flask
openpyxl
pandas
requests
Fabric3
gspread